Update dimension codes
Expand lists in the c_dimcode field where necessary to avoid dropping data when pcornet_demo is populated. Note that the dimcode for American Indian includes both "american ind" and "amerian ind". The latter accomodates a typo currently present in the patient dimension table, to be removed when the source data is fixed.